Made osg::Quat support either float or double internal representation, defaulting to double.
Generalised the osgDB::Field so that its getFloat() method can be used with either doubles or
floats governed by the type passed in - this helps support either float/double
Quat and Matrix classes seemlessly.

+++ src/osgPlugins/osg/Depth.cpp
@@ -57,7 +57,7 @@ bool Depth_readLocalData(Object& obj, Input& fr)
     }
 
     double znear,zfar;
-    if (fr[0].matchWord("range") && fr[1].getDouble(znear) && fr[2].getDouble(zfar))
+    if (fr[0].matchWord("range") && fr[1].getFloat(znear) && fr[2].getFloat(zfar))
     {
         depth.setRange(znear,zfar);
         fr+=2;
